% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genius_artist_songs.R
\name{get_genius_artist_songs}
\alias{get_genius_artist_songs}
\title{Get Artist Songs}
\usage{
get_genius_artist_songs(artist_id,
  access_token = Sys.getenv("GENIUS_API_TOKEN"))
}
\arguments{
\item{artist_id}{Genius Artist ID}

\item{access_token}{Genius Web API token. Defaults to Sys.getenv('GENIUS_API_TOKEN').}
}
\value{
Returns a data frame with the artist's songs information
}
\description{
This function gets all the songs of the artist in genius.
}
\details{
To use this function, you need to first use 'search_genius_artist' to get the artist ID.
}
\examples{
\dontrun{
get_genius_artist_songs('343657')
}
}
\author{
Alberto Almuiña
}
