% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genius_song.R
\name{get_genius_song}
\alias{get_genius_song}
\title{Get Song}
\usage{
get_genius_song(song_id, access_token = Sys.getenv("GENIUS_API_TOKEN"))
}
\arguments{
\item{song_id}{Genius Song ID}

\item{access_token}{Genius Web API token. Defaults to Sys.getenv('GENIUS_API_TOKEN').}
}
\value{
Returns a data frame with the song's information
}
\description{
This function gets 'genius' song information
}
\details{
To use this function, you need to first use 'search_genius_song' to get the song ID.
}
\examples{
\dontrun{
get_genius_song('3440889')
}
}
\author{
Alberto Almuiña
}
