% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\name{get_song}
\alias{get_song}
\title{Get songs.}
\usage{
get_song(song_id, access_token)
}
\arguments{
\item{song_id}{a song's ID on the Genius API}

\item{access_token}{the user's (your) access token to the Genius API.
For more detail about how to get it, see section Details}
}
\value{
A DataFrame with information about a song. Columns:

\describe{
  \item{id}{the song's ID on the Genius API}
  \item{title}{the song's title}
  \item{artist}{the song's main artist}
  \item{featured_artists}{a list of other artists who contributed to the song, if any}
  \item{album}{the song's album}
  \item{date}{the song's released date}
  \item{views}{the number of views for this song on the Genius website}
  \item{contributors}{the number of contributors to the lyrics of this songs on the Genius website}
  \item{transcribers}{the number of transcribers to the lyrics of this songs on the Genius website}
  \item{concurrents}{the number of people looking at the page whenever the API was last updated}
}
}
\description{
This function returns information from the Genius API
about a single song.
}
\examples{
\dontrun{
library(rGenius)

## A Single song ID:
get_song(8439, access_token=YOUR_TOKEN_GOES_HERE)
}

}
\seealso{
\code{\link{get_artist}}, \code{\link{search_song}}, \code{\link{get_songs_from_artist}}
}
