% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\name{get_song_from_artists}
\alias{get_song_from_artists}
\title{Get Songs for an Artist.}
\usage{
get_song_from_artists(artist_name, access_token, n_per_page = 20,
  verbose = FALSE)
}
\arguments{
\item{artist_name}{the name of an artist (will be searched for)}

\item{access_token}{the user's (your) access token to the Genius API.
For more detail about how to get it, see section Details}

\item{n_per_page}{the maximum number of songs that shoud be fetch by the API.
Defaults to 20.}

\item{verbose}{If TRUE, then a progression bar of the execution of the
function is displayed on the console.}
}
\value{
A DataFrame. Columns:

\describe{
  \item{id}{the song's ID on the Genius API}
  \item{title}{the song's title}
  \item{artist}{the song's main artist}
  \item{featured_artists}{a list of other artists who contributed to the song, if any}
  \item{album}{the song's album}
  \item{date}{the song's released date}
  \item{views}{the number of views for this song on the Genius website}
  \item{contributors}{the number of contibutors to the lyrics of this songs on the Genius website}
  \item{transcribers}{the number of transcribers to the lyrics of this songs on the Genius website}
  \item{concurrents}{the number of people looking at the page whenever the API was last updated}
}
}
\description{
This function returns information from the Genius API
about the top songs from an artist, given this artist name.
}
\examples{
\dontrun{
library(rGenius)

## Look up the band Metallica:
get_songs_from_artist("Metallica", access_token=YOUR_TOKEN_GOES_HERE)
}

}
\seealso{
\code{\link{get_artist}}, \code{\link{search_song}}, \code{\link{get_song}}, \code{\link{get_songs}}
}
