% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_song_media.R
\name{get_song_media}
\alias{get_song_media}
\title{Get Song Media}
\usage{
get_song_media(song_id, access_token = Sys.getenv("GENIUS_API_TOKEN"))
}
\arguments{
\item{song_id}{Genius Song ID}

\item{access_token}{Genius Web API token. Defaults to Sys.getenv('GENIUS_API_TOKEN').}
}
\value{
Returns a tibble with the media's information
}
\description{
This function gets 'genius' media information
}
\details{
To use this function, you need to first use 'search_genius_song' to get the song ID.
This function returns the provider (YouTube, Spotify ..), the type (audio, video) and a link for the content.
}
\examples{
\dontrun{
get_song_media('55882')
}
}
\author{
Alberto Almuiña
}
