% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_song_producer_artist.R
\name{get_song_producer_artist}
\alias{get_song_producer_artist}
\title{Get Song Producer}
\usage{
get_song_producer_artist(song_id,
  access_token = Sys.getenv("GENIUS_API_TOKEN"))
}
\arguments{
\item{song_id}{Genius Song ID}

\item{access_token}{Genius Web API token. Defaults to Sys.getenv('GENIUS_API_TOKEN').}
}
\value{
Returns a tibble with the producer's information
}
\description{
This function gets song producer information
}
\details{
To use this function, you need to first use 'search_genius_song' to get the song ID.
This function returns the producer name and links to their genius profiles and images.
}
\examples{
\dontrun{
get_song_producer_artist('55882')
}
}
\author{
Alberto Almuiña
}
