% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_song_relationship.R
\name{get_song_relationship}
\alias{get_song_relationship}
\title{Get Song Relationship}
\usage{
get_song_relationship(song_id,
  access_token = Sys.getenv("GENIUS_API_TOKEN"))
}
\arguments{
\item{song_id}{Genius Song ID}

\item{access_token}{Genius Web API token. Defaults to Sys.getenv('GENIUS_API_TOKEN').}
}
\value{
Returns a tibble with related song's information
}
\description{
This function fetches songs related to the wanted one.
}
\details{
To use this function, you need to first use 'search_genius_song' to get the song ID.
}
\examples{
\dontrun{
get_song_relationship('55882')
}
}
\author{
Alberto Almuiña
}
