% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\name{get_songs}
\alias{get_songs}
\title{Get multiple songs.}
\usage{
get_songs(song_ids, access_token, verbose = TRUE)
}
\arguments{
\item{access_token}{the user's (your) access token to the Genius API.
For more detail about how to get it, see section Details}

\item{verbose}{If TRUE, then a progression bar of the execution of the
function is displayed on the console.}

\item{res}{a res (result) DataFrame. Can be empty (i.e., will be populated here).}

\item{itera}{song ids to iterate over and harvest}
}
\value{
a DataFrame populated with the output of \code{get_song()}
for all songs in song_ids. Columns:

\describe{
  \item{id}{the song's ID on the Genius API}
  \item{title}{the song's title}
  \item{artist}{the song's main artist}
  \item{featured_artists}{a list of other artists who contributed to the song, if any}
  \item{album}{the song's album}
  \item{date}{the song's released date}
  \item{views}{the number of views for this song on the Genius website}
  \item{contributors}{the number of contibutors to the lyrics of this songs on the Genius website}
  \item{transcribers}{the number of transcribers to the lyrics of this songs on the Genius website}
  \item{concurrents}{the number of people looking at the page whenever the API was last updated}
}
}
\description{
harvest information on multiple songs.
}
\examples{
\dontrun{
library(rGenius)

## Multiple song IDs:
get_songs(c(3846, 9869, 2273), access_token=YOUR_TOKEN_GOES_HERE)
}

}
