% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_genius_artist.R
\name{search_genius_artist}
\alias{search_genius_artist}
\title{Search Artist}
\usage{
search_genius_artist(artist_name,
  access_token = Sys.getenv("GENIUS_API_TOKEN"))
}
\arguments{
\item{artist_name}{String of artist name}

\item{access_token}{Genius Web API token. Defaults to Sys.getenv('GENIUS_API_TOKEN').}
}
\value{
Returns a data frame with the artist's information
}
\description{
This function searches 'genius' artists by name
}
\details{
All Genius API endpoints require an artist or song ID. The `search_` functions in this package
try to search for the desired artists/songs to obtain their Genius ID and to use them later.
}
\examples{
\dontrun{
search_genius_artist('Arctic Monkeys')
}
}
\author{
Alberto Almuiña
}
