% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_genius_song.R
\name{search_genius_song}
\alias{search_genius_song}
\title{Search Song}
\usage{
search_genius_song(song_name,
  access_token = Sys.getenv("GENIUS_API_TOKEN"))
}
\arguments{
\item{song_name}{String of song name}

\item{access_token}{Genius Web API token. Defaults to Sys.getenv('GENIUS_API_TOKEN').}
}
\value{
Returns a data frame with the song's information
}
\description{
This function searches 'genius' songs by name
}
\details{
All Genius API endpoints require an artist or song ID. The `search_` functions in this package
try to search for the desired artists/songs to obtain their Genius ID and to use them later.
}
\examples{
\dontrun{
search_genius_song('Paradise')
}
}
\author{
Alberto Almuiña
}
