% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist.R
\name{get_artist}
\alias{get_artist}
\title{Get Artist Information}
\usage{
get_artist(artist_id, access_token)
}
\arguments{
\item{artist_id}{an artist's ID on the Genius API}

\item{access_token}{the user's (your) access token to the Genius API. For more detail about how to get it, see section Details}
}
\value{
a DataFrame of information about an Artist. Columns:

\describe{
 \item{id}{the artist's ID on the Genius API}
 \item{alternate_name}{the artist's name}
 \item{facebook_name}{the artist's name on facebook, if provided on the API}
 \item{instagram_name}{the artist's name on instagram, if provided on the API}
 \item{twitter_name}{the artist's name on twitter, if provided on the API}
 \item{description}{a descriptioon of the artist}
}
}
\description{
This function returns information from the
Genius API about an artist, given this artist's ID on the API.
}
\examples{
\dontrun{
library(rGenius)

## Get Carlos Santana's information
get_artist(8439, access_token=YOUR_TOKEN_GOES_HERE)

}

}
\seealso{
\code{\link{get_song}}, \code{\link{get_songs}}, \code{\link{search_song}}, \code{\link{get_songs_from_artist}}.
}
